pyproject.toml
[project]
name = "asyncapi3"
dynamic = ["version"]
description = "Python AsyncAPI 3.0 object model"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyaml>=25.7.0",
    "pydantic>=2.12.5",
]
license = { text = "MIT" }
authors = [
    { name = "Andrey Shpak" },
]
maintainers = [
    { name = "Andrey Shpak" },
]

[project.urls]
repository = "https://github.com/insspb/asyncapi3"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.sdist]
include = ["asyncapi3"]

[tool.hatch.build.targets.wheel]
include = ["asyncapi3"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cases>=3.9.1",
    "pytest-cov>=7.0.0",
]

[tool.mypy]
plugins = ['pydantic.mypy']
disable_error_code = ["import-untyped"]

[[tool.mypy.overrides]]
module = ["asyncapi3"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disable_error_code = ["attr-defined", "union-attr", "call-arg"]


[tool.pytest.ini_options]
addopts = """
    --junit-xml=./xunit.xml
    --cov=.
    --cov-report=xml
    --cov-report=term-missing:skip-covered
    -vvv
"""
filterwarnings = [
    'ignore:Field name.*shadows.*:UserWarning',
]
junit_duration_report = "total"

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = [
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pyc",
    ".pyi",
    ".tox",
    ".venv",
    "__pycache__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "eggs",
]
[tool.ruff.lint]
# @formatter:off Комментарий необходим, чтобы pycharm не калечил форматирование списка
ignore = [
    "ANN401",  # Dynamically typed expressions (typing.Any) are disallowed in {name}
    "COM812",  # Conflict with RUFF Formatter
    "RUF015",  # Prefer next({iterable}) over single element slice
    "RUF059",  # Unpacked variable {name} is never used
    "SIM105",  # Use contextlib.suppress when possible (Excluded as slow comparable to regular pass)
]

# В пункте select приведён список ВСЕХ поддерживаемых правил в ruff.
# Просьба ограничиться комментированием\расскомментированем необходимых правил, без удаления.
select = [
    "A",        # flake8-builtins
    "AIR",      # Airflow
    "ANN",      # flake8-annotations
#    "ARG",     # flake8-unused-arguments
    "ASYNC",    # flake8-async
    "B",        # flake8-bugbear
    "BLE",      # flake8-blind-except
    "C4",       # flake8-comprehensions
    "C90",      # mccabe
#    "COM",     # flake8-commas
#    "CPY",     # flake8-copyright
#    "D",       # pydocstyle
#    "DJ",      # flake8-django
#    "DOC",     # pydoclint
#    "DTZ",     # flake8-datetimez
    "E",        # pycodestyle errors
#    "EM",      # flake8-errmsg
#    "ERA",     # eradicate
#    "EXE",     # flake8-executable
    "F",        # Pyflakes
#    "FA",      # flake8-future-annotations
#    "FAST",    # FastAPI
#    "FBT",     # flake8-boolean-trap
#    "FIX",     # flake8-fixme
#    "FLY",     # flynt
#    "FURB",    # refurb
#    "G",       # flake8-logging-format
    "I",        # isort
#    "ICN",     # flake8-import-conventions
    "INP",      # flake8-no-pep420
#    "INT",     # flake8-gettext
#    "ISC",     # flake8-implicit-str-concat
#    "LOG",     # flake8-logging
    "N",        # pep8-naming
#    "NPY",     # NumPy-specific rules
#    "PD",      # pandas-vet
#    "PERF",    # Perflint
#    "PGH",     # pygrep-hooks
#    "PIE",     # flake8-pie
#    "PL",      # Pylint
#    "PLC",     # Convention
#    "PLE",     # Error
#    "PLR",     # Refactor
#    "PLW",     # Warning
    "PT",       # flake8-pytest-style
    "PTH",      # flake8-use-pathlib
#    "PYI",     # flake8-pyi
    "Q",        # flake8-quotes
    "RET",      # flake8-return
#    "RSE",     # flake8-raise
    "RUF",      # Ruff-specific rules
    "S",        # flake8-bandit
    "SIM",      # flake8-simplify
#    "SLF",     # flake8-self
#    "SLOT",    # flake8-slots
    "T10",      # flake8-debugger
    "T20",      # flake8-print
#    "TC",      # flake8-type-checking
#    "TD",      # flake8-todos
    "TID",      # flake8-tidy-imports
#    "TRY",     # tryceratops
    "UP",       # pyupgrade
    "W",        # pycodestyle warnings
#    "YTT",     # flake8-2020
]

[tool.ruff.lint.per-file-ignores]
"asyncapi3/models/bindings/__init__.py" = ["RUF022"]
"__init__.py" = [
    "F401",     # unused import
    "F403",     # undefined name
]
"test_*.py" = [
    "S101",     # Use of `assert`
    "S106",     # Possible hardcoded password
    "E501",     # line-too-long
]

[tool.ruff.lint.mccabe]
max-complexity = 12
# @formatter:on

[tool.ruff.lint.isort]
lines-between-types = 1
